Time: 3 ms (25.64%), Space: 5.3 MB (61.54%) - LeetHub

+++ 1344-angle-between-hands-of-a-clock/1344-angle-between-hands-of-a-clock.c
@@ -3,7 +3,7 @@ double angleClock(int hour, int minutes)
     double minDegree = 0, hourDegree = 0, correctTime = 0;
     
     minDegree = minutes * 6;
-    hourDegree = (double)(hour * 30 + (double) ((double)minutes * (double)0.5) );
+    hourDegree = (double)(hour * 30 + (double) (minutes *0.5));
     
     correctTime = (minDegree - hourDegree < 0) ? (hourDegree - minDegree) : (minDegree - hourDegree);
